perf(ui): defer custom theme import to settings
